requirements.txt
click>=2.0
bs4
pyyaml
mechanize
bibtexparser
